add extra authorization for delete brand and property (#357)

+++ backend-node/src/resolvers/property/deletePropertyMutation.ts
@@ -10,9 +10,15 @@ let deletePropertyResolver: FieldResolver<
     where: {
       id: propertyId,
     },
+    include: {
+      landlordUser: true,
+    },
   });
   if (!property) {
     throw new Error('Property not found!');
+  }
+  if (property.landlordUser?.id !== context.landlordUserId) {
+    throw new Error('User not authorized to delete');
   }
   await context.prisma.property.delete({
     where: {
